Added missing Promise.reject
Hi,

In order to get the behavior described in the tutorial:

"This code handles both a successful and unsuccessful login. On success, you call the frame module's `navigate()` method to navigate the user to the (currently empty) list page. On failure, you use the dialog module to show the user an error message. Try inputting some invalid credentials to see what the dialog looks like."

You need to add `return Promise.reject();` in the end of the user.login().catch function. Otherwise the user would still be redirected to the list on login failure.

I looked at the repo for the Groceries app and the code is there.

Great tutorial by the way :)