Update debugging.md
In other Telerik documentation, prior to a trip I made to Bulgaria almost three years ago, we determined that more customers find a given page through search than by navigating through links in Telerik's own docs. Given that will likely be true for the NativeScript docs, I added a comment at the top to provide some "setup" info before you tell the customer how to start debugging. This help article assumes the customer is at a comment prompt and has certain software installed. I think you should provide a link to the page with that info or add it specifically. Also, this help article does not say say anything about using an emulator, nor does it link to a page that explains what it is or how to use it.

adding debug help commands- iOS/android