Update animation.md
Wow, this is probably the longest article I have ever edited for Telerik. :-) I suggest that you break this up into smaller articles. The content is GREAT but many things might be hard to find (using an external search engine) because there should be many keywords used, possibly too many to describe everything in this article. In my other job, our SEO person says 600 words is about the optimum size for an article to be search friendly. If you get longer than 2000 words, the SEO decreases because few people will read past that.
I added many comments to this, including about 21 places to add captions for code examples.

add captions and edit some of the comments

Split the animation article on three pieces