Update upgrade-instructions.md
Mostly minor edits, but before someone installs an upgrade, don't they have to download something? Does the install of the upgrade also download it?  I suggest that you add a comment about that. To me, it felt like the article was missing that step or an explanation of why it isn't necessary.