Update running-latest.md
There are three inline comments for you to address.

I also have a general style question. In both the "Behind the curtains of running a NativeScript application" section and the "Building the repo" section, in the "term" list you used hyphens after the first/defined/emphasized term rather than a colon which is used in other NativeScript term lists. I had a discussion with the XAML team and they are moving from hyphens to colons, and prefer to have the first term bold. Is there a reason you chose to use hyphens here?