Update plugins.md
In the last section, Manual Steps After Removal, there are enough steps for the reader to perform in sequence that I suggest you rewrite this section slightly to make it an ordered list. That will make it easier for the reader to follow every step and not skip any.

Other than that, I made minor copy editing changes. This was a very good article.