Updated @import syntax for theme files
If you are using an older version of the `nativescript-dev-sass` plugin (before 1.0.0), the tilde ("~") in the `@import` is **not** required for referencing theme stylesheets. As of 1.0.0, the "~" is required to tell the node sass compiler to resovle the import file path relative to the `node_modules` directory. Updated docs to show new required syntax.