Update action-bar.md
Website androiddrawables.com is not working. But, I could not find the alternative website. 
The website https://developer.android.com/reference/android/R.drawable.html seemed to me the one which should be followed for the system icons, but this page did not display the icons corresponding to the icon name.