Delete obsolete paragraph
Following the link to the webinar merely produces a page saying the webinar has ended. This makes the paragraph useless. If the webinar were recorded and available, then the paragraph would be useful, but it would need to omit the statement that the user will have an opportunity to ask questions.