Update ns-setup-linux.md
As discussed in [this SO question](https://stackoverflow.com/questions/7535569/what-is-an-appropriate-directory-in-which-to-install-android-sdk). `/usr/local` should be the right place for the android sdk.

Also putting it is the `/` seems to be a bad idea. The android sdk is too large to be put in `/` in my opinion.

Regarding the environmental variable, I user `~/.bashrc` in favor of `/etc/environment`. For some reason using `/etc/environment` won't cut it. 

A [reference](https://stackoverflow.com/questions/26256279/how-to-set-android-home-path-in-ubuntu) for using `~/.bashrc`