Update ns-setup-linux.md
When I tryed to follow step 6 to found usr/lib/android-sdk I have got an error `Not Found`. Problem was in next. We set as ANDROID_HOME variable `usr/lib/android-sdk/` but at step 5 we extract an archive to `usr/local/android/sdk`. As you can see it is different paths.